delete tips

+++ src/views/job-management/job-set.vue
@@ -468,25 +468,40 @@ export default {
     },
     //删除
     deleteMsg() {
+      if (this.multipleSelection.length ==0) {
+          this.$message.error("请选择一条数据");
+          return;
+      }
       if (this.multipleSelection.length > 1) {
         this.$message.error("只能选择一个删除");
         return;
       }
-      this.$api.jobUserManagement
-        .deleteTableData(this.multipleSelection[0].foid)
-        .then((res) => {
-          if (res.data.code == 0) {
-            this.$message.success("删除成功");
-            this.isEdit = false;
-            //刷新表格
-            this.getTableData("");
-          } else {
-            this.$message.error(res.data.msg);
-          }
-        }),
-        (error) => {
-          console.log(error);
-        };
+        this.$confirm('此操作将永久删除, 是否继续?', '提示', {
+            confirmButtonText: '确定',
+            cancelButtonText: '取消',
+            type: 'warning'
+        }).then(() => {
+            this.$api.jobUserManagement
+                .deleteTableData(this.multipleSelection[0].foid)
+                .then((res) => {
+                    if (res.data.code == 0) {
+                        this.$message.success("删除成功");
+                        this.isEdit = false;
+                        //刷新表格
+                        this.getTableData("");
+                    } else {
+                        this.$message.error(res.data.msg);
+                    }
+                }),
+                (error) => {
+                    console.log(error);
+                };
+        }).catch(() => {
+            this.$message({
+                type: 'info',
+                message: '已取消删除'
+            });
+        });
     },
 
     // 编辑
